Refactor : 스터디 관련 컴포넌트 수정
